+++ src/coreclr/gc/gcimpl.h
@@ -332,7 +332,7 @@ public:
 
     virtual int RefreshMemoryLimit();
 
-    virtual void NullBridgeObjectsWeakRefs(int length, void* unreachableObjectHandles);
+    virtual void NullBridgeObjectsWeakRefs(size_t length, void* unreachableObjectHandles);
 };
 
 #endif  // GCIMPL_H_
